HACK Add initial rofi processor
Follows my personal rofi setup, which is presumably non-default. Puts
the line including colorscheme into `.config/rofi/themes/settings.rasi`,
I don't remember if that is default. TODO

Also uses base16 templates from the `themes` folder of the package,
which set some *layout* properties of rofi not just color changes.
Should probably look at rofi and create a 'nullifying' rofi rasi file
similar to css plate cleansers and build a default layout from that.

Other way is to use `colors` subfolder of the package, but this does not
connect the colors to uses in the themes, so would be more work to
include.