Add experimental dunst support
Added simple support for dunst files. Since files can, similar to
alacritty, not be included in dunstrc, the processor invokes a similarly
invasive replacement method to add the base16 at the end of the config
file.

The procedure is similar to alacritty processor, the danger is also the
same -- it works directly on the config file and will worst case delete
the complete contents of this file. Since the processor is not tested
(other than manually looking that it does what it's supposed to) the
case of this happening is not negligible.