Fix polybar hardcoding
Polybar restarting is now done through its provided ipc interface,
instead of relying on a hardcoded script on the system.

Previously, the system needed access to a `polybar-launch` script.
Now, it invokes the `polybar-msg` command which is provided by polybar
itself.

It is important that polybar has ipc enabled in its configuration for
the restart to work.